Add further dtype tests

+++ tests/fingerprints/avalon.py
@@ -8,7 +8,7 @@ from skfp.fingerprints import AvalonFingerprint
 def test_avalon_bit_fingerprint(smiles_list, mols_list):
     avalon_fp = AvalonFingerprint(sparse=False, count=False, n_jobs=-1)
     X_skfp = avalon_fp.transform(smiles_list)
-    X_rdkit = np.array([GetAvalonFP(mol) for mol in mols_list], dtype=np.uint8)
+    X_rdkit = np.array([GetAvalonFP(mol) for mol in mols_list])
 
     assert np.array_equal(X_skfp, X_rdkit)
     assert X_skfp.shape == (len(smiles_list), avalon_fp.fp_size)
@@ -19,9 +19,7 @@ def test_avalon_bit_fingerprint(smiles_list, mols_list):
 def test_avalon_count_fingerprint(smiles_list, mols_list):
     avalon_fp = AvalonFingerprint(sparse=False, count=True, n_jobs=-1)
     X_skfp = avalon_fp.transform(smiles_list)
-    X_rdkit = np.array(
-        [GetAvalonCountFP(mol).ToList() for mol in mols_list], dtype=np.uint32
-    )
+    X_rdkit = np.array([GetAvalonCountFP(mol).ToList() for mol in mols_list])
 
     assert np.array_equal(X_skfp, X_rdkit)
     assert X_skfp.shape == (len(smiles_list), avalon_fp.fp_size)
@@ -32,7 +30,7 @@ def test_avalon_count_fingerprint(smiles_list, mols_list):
 def test_avalon_sparse_bit_fingerprint(smiles_list, mols_list):
     avalon_fp = AvalonFingerprint(sparse=True, count=False, n_jobs=-1)
     X_skfp = avalon_fp.transform(smiles_list)
-    X_rdkit = csr_array([GetAvalonFP(mol) for mol in mols_list], dtype=np.uint8)
+    X_rdkit = csr_array([GetAvalonFP(mol) for mol in mols_list])
 
     assert np.array_equal(X_skfp.data, X_rdkit.data)
     assert X_skfp.shape == (len(smiles_list), avalon_fp.fp_size)
@@ -43,9 +41,7 @@ def test_avalon_sparse_bit_fingerprint(smiles_list, mols_list):
 def test_avalon_sparse_count_fingerprint(smiles_list, mols_list):
     avalon_fp = AvalonFingerprint(sparse=True, count=True, n_jobs=-1)
     X_skfp = avalon_fp.transform(smiles_list)
-    X_rdkit = csr_array(
-        [GetAvalonCountFP(mol).ToList() for mol in mols_list], dtype=np.uint32
-    )
+    X_rdkit = csr_array([GetAvalonCountFP(mol).ToList() for mol in mols_list])
 
     assert np.array_equal(X_skfp.data, X_rdkit.data)
     assert X_skfp.shape == (len(smiles_list), avalon_fp.fp_size)
